Share pipeline when possible
Even though we share credentials and transports, there are instances
where it is beneficial to share the entire pipeline, especially for
datasets created using from_blob_urls().

Specifically, pipeline policy instances may cache resources that
dramatically improve performance and policy instances are tightly
attached to pipeline instances.

For example, the bearer token credential policy caches tokens
retrieved from a credential object. However, policy instances are
only shared if transports are shared. So for cases where a client is
not part of the same heirarchy (e.g. clients from BlobDataset.from_blob_urls),
retrieved, still-valid tokens are not reused. As a result, if
DefaultAzureCredential resolves to a provider that does not cache
credentials at all (e.g. AZ CLI credentials), we can end up making an
expensive call (e.g. invoke the CLI) for every client that is created
that may be just used for a single GET request. So by reusing the
pipeline we reuse the cached token on the policy instance.

This improved loading a dataset from_blob_urls() by orders of magnitudes
and is now comparable to if we provided SAS tokens, and not require
a get_token() call.

The one caveat is that we cannot always share pipelines if a resource
url includes a SAS override. We need a separate pipeline for that client
as the credential policy is coupled to a pipeline. This is not as much
of an issue because we are still sharing the transport layer and if we
are using SAS, credential caching does not apply.