Plumb connection data block size to transport
Previously this had been provided to the client's constructor.
However, because we are now creating the transport ourselves and
sharing it as part of the factory, the configuration is no longer
used when provided to the client since we are providing an
explicit transport. Instead, we need to provide it to the transport
we create.